(!) Major refactor/reformat and overhauling
Made broad, sweeping changes to try to speed up on things and hurry the "Great Migration" to nullability correctness and code cleanliness. This branch has some problems/bugs, but still works well and is usable. The issues in it are manageable and can be fixed with some debugging and testing things.